Modified routes and apis and cleaned code
