+++ root/app/server.js
@@ -1,20 +1,15 @@
 console.log('Starting Server Init')
 
 var express           = require('express')
-var routes            = require('../routes/index')
-var user_routes       = require('../routes/user')
 var http              = require('http')
 var path              = require('path');
-var app               = express();
 var bodyParser        = require('body-parser');
 var cookieParser      = require('cookie-parser');
 var morgan            = require('morgan');
 var passport          = require('passport');
-var config            = require('../config/database');
-var User              = require('./models/user');
 var jwt               = require('jwt-simple');
 var app_middleware    = require('./middleware')
-var require_login     = app_middleware.require_login
+var app               = express();
 
 app.set('views', __dirname + '/../views');
 app.set('view engine', 'jade');
@@ -44,21 +39,10 @@ if ('development' == app.get('env')) {
 app.use(app_middleware.get_user);
 app.use(app.router);
 
-// URLs
-require('../routes/game_and_chat.js')(app);
-require('../routes/comments.js')(app );
-app.get('/instructions', routes.instructions);
-app.get('/theme', routes.theme);
-app.get('/author', routes.author);
-app.get('/register', user_routes.register);
-app.get('/login', user_routes.login);
-app.get('/logout', user_routes.logout);
-app.get('/leaderboard', require_login, routes.leaderboard);
-
-// API
-app.post('/api/signup', user_routes.signup);
-app.post('/api/authenticate', user_routes.authenticate);
-app.post('/api/set-high-score', passport.authenticate('jwt', { session: false}), user_routes.set_high_score);
-app.get('/api/get-high-score', passport.authenticate('jwt', { session: false}), user_routes.get_high_score);
+//Routes
+require('../routes/index.js')(app);
+require('../routes/api/chats.js')(app);
+require('../routes/api/comments.js')(app );
+require('../routes/api/user.js')(app );
 
 var server = module.exports = http.createServer(app);
