+++ root/routes/index.js
@@ -1,30 +1,79 @@
-User = require('../app/models/user');
+var User           = require('../app/models/user');
+var app_middleware = require('../app/middleware')
+var require_login  = app_middleware.require_login
+const fs           = require('fs');
+const path         = require('path');
 
-exports.instructions = function(req, res){
-  res.render('instructions', {
-    title: "Instructions",
-  })
-};
+function renderGamePage(res) {
+   var images = fs.readdirSync(path.join(__dirname, '../static/images/game')).map(function(image){
+      return image.substring(0, image.length - 4);
+   })
+   res.render('game', {title : "Play Action Box",images : images})
+}
 
-exports.theme = function(req, res){
-  res.render('theme', {
-    title: "Theme",
-  })
-};
+module.exports = function (app){
 
-exports.author = function(req, res){
-  res.render('author', {
-    title: "Authors",
-  })
-};
+  app.get('/', require_login, function (req, res) {
+    renderGamePage(res);
+  });
 
-exports.leaderboard = function(req, res){
-  User.find().sort({ highscore: 'descending' }).exec(function(err, users) {
+  app.get('/play', require_login, function (req, res) {
+    renderGamePage(res);
+  });
+
+  app.get('/instructions', function(req, res){
+    res.render('instructions', {
+      title: "Instructions",
+    })
+  });
+
+  app.get('/theme', function(req, res){
+    res.render('theme', {
+      title: "Theme",
+    })
+  });
+
+  app.get('/author', function(req, res){
+    res.render('author', {
+      title: "Authors",
+    })
+  });
+
+  app.get('/comments', require_login, function (req, res) {
+    res.render('comments',{title: "Comments"});
+  });
+
+  app.get('/register',  function(req, res){
+   context = {
+      title: "Register",
+    };
+    res.render('register', context);
+  });
+
+  app.get('/login', function(req, res){
+    res.cookie('auth_token' , 'undefined');
     context = {
-      title: "Leaderboard",
-      users: users.slice(0,10),
-    }
-    res.render('leaderboard', context);
-  })
+      title: "Login",
+      perm_denied: req.param('perm_denied', false),
+    };
+    res.render('login', context);
+  });
 
+  app.get('/logout', function(req, res){
+    res.cookie('auth_token' , 'undefined');
+    res.render('logout', {
+      title: 'Logout',
+      user: undefined,
+    });
+  });
+
+  app.get('/leaderboard', function(req, res){
+    User.find().sort({ highscore: 'descending' }).exec(function(err, users) {
+      context = {
+        title: "Leaderboard",
+        users: users.slice(0,10),
+      }
+      res.render('leaderboard', context);
+    })
+  });
 }
